Major update to the model build process which includes rework to support he status bar on svn checkout, background kext fetching with status and many improvements to the build process.

+++ trunk/bin/html/workerapp.php
@@ -281,7 +281,7 @@ function fixes_reset_display() {
 }
 
 function showBuildLog() {
-	global $workpath;
+	global $workpath, $edp, $ee;
 	echoPageItemTOP("icons/big/logs.png", "Build configuration log");
 	echo "<body onload=\"JavaScript:timedRefresh(5000);\">";	
 
@@ -292,17 +292,86 @@ function showBuildLog() {
 	echo "<b><center>Please wait for few minutes while we download the files needed for your model (will take approx 5 - 15 minutes depending on your internet speed) and it will start building after the download................</center></b><br>";*/
 	
 	echo "<b>Build process:</b><br>";
-	include "$workpath/build.log";
+	
+	if(is_file("$workpath/build.log"))
+		include "$workpath/build.log";
+		
+	if(is_file("$workpath/myFix.log"))
+		include "$workpath/myFix.log";
+		
+	if(is_file("$workpath/myFix2.log"))
+		include "$workpath/myFix2.log";
+	
+	if(is_dir("$workpath/kpsvn/dload/statFiles")) {
+		$fcount = shell_exec("cd $workpath/kpsvn/dload/statFiles; ls | wc -l");
+	}
+	if ($fcount > 0)
+		echo "<b>Files left to download/update : $fcount</b><br>";
+		
+	if ($fcount == 0 && !is_file("$workpath/myFix.log"))
+	{
+		
+		$edp->writeToLog("$workpath/build.log", "<br><b>All Files downloaded/updated.</b><br>");
+		
+		//
+		// Step 5 : Copying custom files from /Extra/include
+		//
+		$edp->writeToLog("$workpath/build.log", "<br><b>Step 5) Copying custom files from /Extra/include:</b><br>");
+		copyCustomFiles();
+		
+		//
+		// Step 6 : Applying last minute fixes and generating caches
+		//
+		$edp->writeToLog("$workpath/myFix.log", "<br><b>Step 6) Applying last minute fixes and Calling myFix to copy kexts & generate kernelcache:</b><br>");
+		
+		// Final Chown to SLE and touch (this is due to some issuses with myFix in Mavericks)
+		system_call("sudo chown -R root:wheel /System/Library/Extensions/");
+		system_call("sudo touch /System/Library/Extensions/");
+		$edp->writeToLog("$workpath/myFix.log", "Clearing boot-args in NVRAM...<br>");
+		system_call("nvram -d boot-args");
+		$edp->writeToLog("$workpath/myFix.log", "Removing version control of kexts in $ee<br>");
+   		system_call("rm -Rf `find -f path \"$ee\" -type d -name .svn`");
+   		
+   		//$edp->writeToLog("$workpath/kpsvn/dload/myFix.sh", "sudo myfix -q -t / >> $workpath/myFix.log &");
+		$edp->writeToLog("$workpath/myFix.log", "<a name='myfix'></a>");
+		$edp->writeToLog("$workpath/myFix.log", "Running myFix to fix permissions and genrate cache...<br><br>");
+		$edp->writeToLog("$workpath/myFix.log", "<b>* * * * * * * * * * * *  myFix process status * * * * * * * * * * * *</b><br><pre>");
 
+		//shell_exec("sudo myfix -q -t / >> $workpath/myFix.log &");
+		/*$edp->writeToLog("$workpath/kpsvn/dload/myFix.sh", "sudo myfix -q -t / >> $workpath/myFix.log &");
+		system_call("sh $workpath/kpsvn/dload/myFix.sh &");*/
+	}
 	echo "<script type=\"text/JavaScript\"> function timedRefresh(timeoutPeriod) { logVar = setTimeout(\"location.reload(true);\",timeoutPeriod); } function stopRefresh() { clearTimeout(logVar); } </script>\n";
 	echo "</div>";
 }
 
 function showLoadingLog() {
-	global $workpath;
-	echo "<div class='pageitem_bottom'\">";
-	echo "<b><center>After starting the build process, please wait for few minutes while we download the files needed for your model (will take approx 5 - 15 minutes depending on your internet speed). You will be redirected to the build process log at the end of the build.</center></b>";
-	echo "<img src=\"icons/big/loading.gif\" style=\"width:200px;height:30px;position:relative;left:50%;top:50%;margin:15px 0 0 -100px;\">";
+	global $workpath, $edp, $ee;	
+
+	echo "<div class='pageitem_bottom'\">";	
+	if(is_dir("$workpath/kpsvn/dload/statFiles")) {
+		$fcount = shell_exec("cd $workpath/kpsvn/dload/statFiles; ls | wc -l");
+	}
+	
+	if ($fcount == "" || $fcount > 0 || !is_file("$workpath/myFix.log")) {
+		echo "<body onload=\"JavaScript:timedRefresh(8000);\">";
+		echo "<center><b>After starting the build process, please wait for few minutes while we download the files needed for your model.</b> [which will take approx 5 to 15 minutes depending on your internet speed] <br><br><b>Shortly you will be redirected to the build process log which will show the status of the build.</center></b>";
+		echo "<img src=\"icons/big/loading.gif\" style=\"width:200px;height:30px;position:relative;left:50%;top:50%;margin:15px 0 0 -100px;\">";
+		
+		if ($fcount > 0)
+			echo "<br><b> Files left to download/update : $fcount</b><br>";
+		
+		echo "<script type=\"text/JavaScript\"> function timedRefresh(timeoutPeriod) { logVar = setTimeout(\"location.reload(true);\",timeoutPeriod); } function stopRefresh() { clearTimeout(logVar); } </script>\n";
+	}
+	else {
+		if ($fcount == 0 && is_file("$workpath/myFix.log") && !is_file("$workpath/myFix2.log"))
+		{
+			shell_exec("sudo myfix -q -t / >> $workpath/myFix2.log &");
+			//system_call("sudo myfix -q -t / >> $workpath/myFix.log &");
+		}
+		echo "<img src=\"icons/big/success.png\" style=\"width:80px;height:80px;position:relative;left:50%;top:50%;margin:15px 0 0 -35px;\">";
+		echo "<b><center> Build Finished, Please wait for the myFix process to finish fixing permissions and generating caches.</b> [check the build log on right side for status] <br><br><b> You can then reboot your system (or) close this app.</center></b>";
+	}	
 	echo "</div>";
 }
